Synchronize velocity for diagnostics (#1751)
This PR allows the synchronization in time of the particle velocities
and positions when generating diagnostics. Without this option, the
particle velocities will lag behind the position by a half time step.
This adds the boolean input parameter
`warpx.synchronize_velocity_for_diagnostics` to turn on this option,
defaulting to false.

There are several pieces to this PR:
- Changes to `MultiDiagnostic` and `MultiReducedDiags` adding routines
to check if any diagnostics will be done
- Adds a call to `PushP` to just before the diagnostics are done (to get
the updated fields from the electrostatic calculation)
- Add the appropriate documentation

What `Evolve` does is if the synchronization is to be done, advance the
velocity a half step just before the diagnostics and sets
`is_synchronized=true`. Then at the start of the next step, if
`is_synchronized` is true, push the velocities back a half step to be
ready for the full leap frog advance.

Comments:
- Is the documentation in the correct place in parameters.rst?
- The reduced diagnostics could perhaps use the new DoDiags method
instead of accessing `m_intervals` in its ComputeDiags.
- This PR leaves the original PushP unchanged, even though it is
possibly buggy. That PushP fetches the fields, but uses the particle
positions before the particle boundary conditions have been applied,
leading to a possible out of bounds reference. Also, that PushP may not
be consistent with the backwards PushP since the fields may had changed.
Comments are added to the code to note this potential problem. I avoided
changing this since it breaks many CI tests.

---------

Co-authored-by: Edoardo Zoni <ezoni@lbl.gov>
Co-authored-by: Axel Huebl <axel.huebl@plasma.ninja>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Edoardo Zoni <59625522+EZoni@users.noreply.github.com>