CI: Ubuntu 22.04/24.04 in GitHub Actions, CUDA 11.7+ (#5731)
The Ubuntu 20.04 runner image will be deprecated soon:
https://github.com/actions/runner-images/issues/11101.
Only support CUDA 11.7+ going forward.

To-do:
- [x] GitHub Actions workflows
- [ ] ~~Azure DevOps workflows~~ (transfered to #5661)

### Notes
I renamed the following workflows, with the aim of not having to rename
them again in the future when we upgrade version numbers:
- `NVCC 11.3 SP`, now `NVCC SP`
- `NVCC 11.8.0 GNUmake`, now `NVC GNU Make`
- `NVHPC@24.1 NVCC/NVC++ Release [tests]`, now `NVHPC`

If you are okay with the renaming, the list of "required" workflows
should be updated to reflect the new names, and I don't have the
necessary access to do so.